+++ BlueprintUI/Sources/Internal/Extensions/CGRect.swift
@@ -15,12 +15,26 @@ extension CGRect {
     ///   - rule: The rounding rule.
     ///   - scale: The rounding scale.
     /// - Returns: A rectangle with the rounded values.
-    func rounded(_ rule: FloatingPointRoundingRule, by scale: CGFloat) -> CGRect {
+    func roundedPrioritizingEdges(_ rule: FloatingPointRoundingRule, by scale: CGFloat) -> CGRect {
         CGRect(
             minX: minX.rounded(rule, by: scale),
             minY: minY.rounded(rule, by: scale),
             maxX: maxX.rounded(rule, by: scale),
             maxY: maxY.rounded(rule, by: scale)
+        )
+    }
+
+    /// Creates a new rectangle by rounding the origin and size of this rectangle.
+    /// - Parameters:
+    ///   - rule: The rounding rule.
+    ///   - scale: The rounding scale.
+    /// - Returns: A rectangle with the rounded values.
+    func roundedPrioritizingSize(_ rule: FloatingPointRoundingRule, by scale: CGFloat) -> CGRect {
+        CGRect(
+            x: origin.x.rounded(rule, by: scale),
+            y: origin.y.rounded(rule, by: scale),
+            width: size.width.rounded(rule, by: scale),
+            height: size.height.rounded(rule, by: scale)
         )
     }