+++ BlueprintUI/Tests/BlueprintViewTests.swift
@@ -470,6 +470,56 @@ class BlueprintViewTests: XCTestCase {
 
         XCTAssertEqual(events, expectedEvents)
     }
+
+    func test_onAppearAfterViewUpdates() {
+        let childViewDescriptionApplied = expectation(description: "child view description applied")
+        let parentAppeared = expectation(description: "parent appeared")
+
+        let element = LifecycleTestElement(
+            onAppear: {
+                parentAppeared.fulfill()
+            },
+            onDisappear: {},
+            wrapped: CallbackElement(onViewDescriptionApplied: {
+                childViewDescriptionApplied.fulfill()
+            })
+        )
+
+        let view = BlueprintView()
+        let window = UIWindow(frame: UIScreen.main.bounds)
+        window.addSubview(view)
+
+        view.element = element
+        view.ensureLayoutPass()
+
+        // Check that parent element lifecycle callbacks happen *after* view updates
+        // of their children, so things like focus triggers can be set up.
+        wait(for: [childViewDescriptionApplied, parentAppeared], timeout: 1, enforceOrder: true)
+    }
+
+    func test_intrinsicContentSize_noElement() {
+        let view = BlueprintView()
+
+        let noIntrinsicMetricSize = CGSize(
+            width: UIView.noIntrinsicMetric,
+            height: UIView.noIntrinsicMetric
+        )
+
+        XCTAssertEqual(view.intrinsicContentSize, noIntrinsicMetricSize)
+    }
+
+    func test_intrinsicContentSize_constrained() {
+        let view = BlueprintView()
+
+        let size = CGSize(
+            width: 42,
+            height: 7
+        )
+
+        view.element = Empty().constrainedTo(size: size)
+
+        XCTAssertEqual(view.intrinsicContentSize, size)
+    }
 }
 
 fileprivate struct MeasurableElement: Element {
@@ -553,6 +603,22 @@ private struct TestContainer: Element {
 
         func layout(size: CGSize, items: [(traits: (), content: Measurable)]) -> [LayoutAttributes] {
             Array(repeating: LayoutAttributes(size: .zero), count: items.count)
+        }
+    }
+}
+
+private struct CallbackElement: Element {
+    var onViewDescriptionApplied: () -> Void
+
+    var content: ElementContent {
+        ElementContent(intrinsicSize: .zero)
+    }
+
+    func backingViewDescription(with context: ViewDescriptionContext) -> ViewDescription? {
+        UIView.describe { config in
+            config.apply { _ in
+                onViewDescriptionApplied()
+            }
         }
     }
 }