+++ CHANGELOG.md
@@ -11,9 +11,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- The `@FocusState` property wrapper can be used to manage focus of text fields. ([#259])
+
+  After binding a text field to a state, you can programmatically focus the field by setting the state value.
+  
+  ```swift
+  struct LoginForm: ProxyElement {
+      enum Field: Hashable {
+          case username, password
+      }
+
+      @FocusState private var focusedField: Field?
+
+      var elementRepresentation: Element {
+          // This text field will be focused when `self.focusedField = .username`
+          TextField(text: "")
+              .focused(when: $focusedField, equals: .username)
+      }
+  }
+  ```
+
 ### Removed
 
 ### Changed
+
+- `BlueprintView`'s `intrinsicContentSize` will now return `UIView.noIntrinsicMetric` if there is no `element` associated with it.
+
+- `TextField`'s `becomeActiveTrigger` and `resignActiveTrigger` properties have been replaced with a `focusBinding` for use with the new `@FocusState` property wrapper.
 
 ### Deprecated
 
@@ -24,6 +48,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Misc
 
 # Past Releases
+
+## [0.28.1] - 2021-09-10
+
+### Added
+
+- View-backed elements may opt-in to a frame rounding behavior that prioritizes preserving the frame size rather than the frame edges. This is primarily meant for text labels, to fix an issue where labels gain or lose a pixel in rounding and become wrapped or truncated incorrectly. ([#257])
+
+### Changed
+
+- Lifecycle hooks are guaranteed to run after all views are updated. ([#260])
 
 ## [0.28.0] - 2021-09-01
 
@@ -606,7 +640,8 @@ searchField
 
 - First stable release.
 
-[main]: https://github.com/square/Blueprint/compare/0.28.0...HEAD
+[main]: https://github.com/square/Blueprint/compare/0.28.1...HEAD
+[0.28.1]: https://github.com/square/Blueprint/compare/0.28.0...0.28.1
 [0.28.0]: https://github.com/square/Blueprint/compare/0.27.0...0.28.0
 [0.27.0]: https://github.com/square/Blueprint/compare/0.26.0...0.27.0
 [0.26.0]: https://github.com/square/Blueprint/compare/0.25.0...0.26.0
@@ -643,6 +678,9 @@ searchField
 [0.3.1]: https://github.com/square/Blueprint/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/square/Blueprint/compare/0.2.2...0.3.0
 [0.2.2]: https://github.com/square/Blueprint/releases/tag/0.2.2
+[#260]: https://github.com/square/Blueprint/pull/260
+[#259]: https://github.com/square/Blueprint/pull/259
+[#257]: https://github.com/square/Blueprint/pull/257
 [#244]: https://github.com/square/Blueprint/pull/244
 [#209]: https://github.com/square/Blueprint/pull/209
 [#176]: https://github.com/square/Blueprint/pull/176